requirements.txt
django
django-rest-framework
xmltodict
djangorestframework-xml